Update README.md
The function of the PCIe bridge module is to convert the PCIe interface into an AXI interface and an AXI-Lite interface, which is implemented by the XDMA IP provided by Xilinx. The AXI interface enables high-speed transmission of large amounts of data. The AXI-Lite interface enables the low-speed transmission of small amounts of data. In addition, the PCIe bridge module can also receive interrupts sent to the host by other modules in the system, convert them into PCIe interrupts, and feedback the status to the host.